+++ R/zzz.R
@@ -31,7 +31,8 @@ opt.orgdata <- list(
   ## Use for debuging purposes only
   orgdata.debug = FALSE,
   orgdata.debug.nrow = 0,  #If other than 0 then x rows will be selected
-  orgdata.debug.aggregate = FALSE #Recode but don't aggregate if TRUE
+  orgdata.debug.aggregate = FALSE, #Recode but don't aggregate if TRUE
+  orgdata.debug.geo = FALSE #Keep old geo if TRUE
 )
 
 .onLoad <- function(libname, pkgname) {
@@ -42,5 +43,5 @@ opt.orgdata <- list(
 }
 
 .onAttach <- function(libname, pkgname) {
-  packageStartupMessage("orgdata version 0.3.9")
+  packageStartupMessage("orgdata version 0.3.10")
 }